refactor: separate codebase - hash
